+++ src/cli/release/download/download.js
@@ -2,7 +2,14 @@
 const fs = require('fs');
 const chalk = require('chalk');
 const emoji = require('node-emoji');
-const { getPlatform, print, printError, printInfo } = require('core/util');
+const {
+  getForgeDistribution,
+  getOsAsync,
+  print,
+  printError,
+  printInfo,
+  warningUnSupportedOS,
+} = require('core/util');
 const { createAsset, download, formatVersion, DOWNLOAD_FLAGS } = require('./libs/index');
 
 const { DEFAULT_MIRROR, RELEASE_ASSETS } = require('../../../constant');
@@ -10,8 +17,10 @@ const { DEFAULT_MIRROR, RELEASE_ASSETS } = require('../../../constant');
 // eslint-disable-next-line consistent-return
 async function main({ args: [userVersion], opts: { mirror = DEFAULT_MIRROR, releaseDir, force } }) {
   try {
-    const platform = await getPlatform();
-    printInfo(`Detected platform is: ${platform}`);
+    const platform = await getForgeDistribution();
+    const osInfo = await getOsAsync();
+    printInfo(`Detected platform is: ${osInfo.dist}`);
+    warningUnSupportedOS(osInfo.dist);
 
     if (releaseDir && fs.existsSync(releaseDir)) {
       printInfo(`${chalk.yellow(`Using local releaseDir: ${releaseDir}.`)}`);